feat: 1278 bo agrement documents tabs
