+++ packages/backend/src/__tests__/helper/fixtures/agrementsHelper.ts
@@ -1,5 +1,6 @@
 import type { AgrementDto } from "@vao/shared-bridge";
 
+import { AgrementService as AgrementServiceShared } from "../../../shared/agrements/agrements.service";
 import { AgrementService } from "../../../usagers/agrements/agrements.service";
 
 export const createAgrement = async ({
@@ -21,7 +22,7 @@ export const createAgrement = async ({
 };
 
 export const getAgrement = async (agrementId: number) => {
-  const agrement = await AgrementService.get({
+  const agrement = await AgrementServiceShared.getById({
     agrementId,
     withDetails: true,
   });
