+++ packages/backend/src/admin/agrements/agrements.controller.ts
@@ -6,6 +6,7 @@ import {
 } from "@vao/shared-bridge";
 import type { NextFunction } from "express";
 
+import { AgrementService as AgrementServiceShared } from "../../shared/agrements/agrements.service";
 import type { RouteRequest, RouteResponse } from "../../types/request";
 import AppError from "../../utils/error";
 import logger from "../../utils/logger";
@@ -53,8 +54,10 @@ export const AgrementController = {
   ) {
     const agrementId = Number(req.validatedParams!.agrementId);
     try {
-      const agrement: AgrementDto | null =
-        await AgrementService.getById(agrementId);
+      const agrement: AgrementDto | null = await AgrementServiceShared.getById({
+        agrementId,
+        withDetails: true,
+      });
       if (!agrement) throw new AppError("NotFound", { statusCode: 404 });
       res.json({ agrement });
     } catch (error) {
