+++ packages/backend/src/admin/agrements/agrements.service.ts
@@ -12,6 +12,7 @@ import Region from "../../services/geo/Region";
 import { mailService } from "../../services/mail";
 import { getOne as serviceOrganismeGetOne } from "../../services/Organisme";
 import TerritoireService from "../../services/Territoire";
+import { AgrementsRepository as AgrementsRepositoryShared } from "../../shared/agrements/agrements.repository";
 import { AgrementMailUsagers } from "../../usagers/agrements/agrements.mail";
 import AppError from "../../utils/error";
 import logger from "../../utils/logger";
@@ -21,9 +22,6 @@ import { AgrementsRepository } from "./agrements.repository";
 const log = logger(module.filename);
 
 export const AgrementService = {
-  async getById(agrementId: number) {
-    return AgrementsRepository.getById(agrementId);
-  },
   async getHistory(agrementId: number) {
     const history = await AgrementsRepository.getHistory(agrementId);
     return history;
@@ -148,7 +146,10 @@ export const AgrementService = {
     commentaire?: string;
     territoireCode: string;
   }): Promise<boolean> {
-    const agrement = await AgrementsRepository.getById(agrementId);
+    const agrement = await AgrementsRepositoryShared.getById({
+      agrementId,
+      withDetails: false,
+    });
     if (!agrement) {
       log.w("Agrement non trouvé", agrementId);
       throw new FunctionalException(FUNCTIONAL_ERRORS.AGREMENT_NOT_FOUND);
