+++ packages/backend/src/usagers/agrements/agrements.controller.ts
@@ -6,6 +6,7 @@ import {
 } from "@vao/shared-bridge";
 import type { NextFunction } from "express";
 
+import { AgrementService as AgrementServiceShared } from "../../shared/agrements/agrements.service";
 import type { RouteRequest, RouteResponse } from "../../types/request";
 import AppError from "../../utils/error";
 import logger from "../../utils/logger";
@@ -74,7 +75,7 @@ export const AgrementController = {
     }
 
     try {
-      const agrement: AgrementDto | null = await AgrementService.get({
+      const agrement: AgrementDto | null = await AgrementServiceShared.getById({
         agrementId,
         withDetails: true,
       });
